create base class for products
