+++ tbsim/interventions/__init__.py
@@ -10,5 +10,6 @@ from .beta import *
 from .tpt import *
 from .bcg import *
 from .health_seeking import *
+from .products import *
 from .diagnostics import *
 from .treatments import *
